metrics.py
"""
评估指标模块
实现PSNR、SSIM和信息熵计算
"""
import numpy as np
try:
    from skimage.metrics import peak_signal_noise_ratio, structural_similarity
except ImportError:
    # 如果skimage不可用，提供fallback
    print("Warning: scikit-image not available, some functions may not work")
    peak_signal_noise_ratio = None
    structural_similarity = None


def calculate_psnr(img1, img2, data_range=255.0):
    """
    计算PSNR (Peak Signal-to-Noise Ratio)
    
    Args:
        img1: numpy array, 参考图像
        img2: numpy array, 增强后的图像
        data_range: float, 图像数据的范围（默认255）
        
    Returns:
        psnr: float, PSNR值
    """
    # 确保数据类型一致
    img1 = img1.astype(np.float64)
    img2 = img2.astype(np.float64)
    
    # 如果图像值在[0, 1]范围，转换到[0, 255]
    if img1.max() <= 1.0:
        img1 = img1 * 255.0
        img2 = img2 * 255.0
        data_range = 255.0
    
    # 计算PSNR
    if peak_signal_noise_ratio is not None:
        psnr = peak_signal_noise_ratio(img1, img2, data_range=data_range)
    else:
        # Fallback implementation
        mse = np.mean((img1 - img2) ** 2)
        if mse == 0:
            return float('inf')
        psnr = 20 * np.log10(data_range / np.sqrt(mse))
    
    return psnr


def calculate_ssim(img1, img2, data_range=255.0):
    """
    计算SSIM (Structural Similarity Index)
    
    Args:
        img1: numpy array, 参考图像
        img2: numpy array, 增强后的图像
        data_range: float, 图像数据的范围（默认255）
        
    Returns:
        ssim: float, SSIM值
    """
    # 确保数据类型一致
    img1 = img1.astype(np.float64)
    img2 = img2.astype(np.float64)
    
    # 如果图像值在[0, 1]范围，转换到[0, 255]
    if img1.max() <= 1.0:
        img1 = img1 * 255.0
        img2 = img2 * 255.0
        data_range = 255.0
    
    if structural_similarity is not None:
        # 如果是多通道图像
        if len(img1.shape) == 3:
            # 计算每个通道的SSIM并取平均
            ssim_channels = []
            for i in range(img1.shape[2]):
                ssim_channel = structural_similarity(
                    img1[:, :, i], 
                    img2[:, :, i], 
                    data_range=data_range
                )
                ssim_channels.append(ssim_channel)
            ssim = np.mean(ssim_channels)
        else:
            # 灰度图
            ssim = structural_similarity(img1, img2, data_range=data_range)
    else:
        # Fallback: 简化的SSIM计算
        # 这里返回一个近似值，建议安装scikit-image以获得准确结果
        print("Warning: Using simplified SSIM calculation. Install scikit-image for accurate results.")
        ssim = 0.5  # 占位值
        # 可以在这里实现简化的SSIM计算
    
    return ssim


def calculate_entropy(image):
    """
    计算图像的信息熵
    
    Args:
        image: numpy array, 输入图像
        
    Returns:
        entropy_value: float, 信息熵值
    """
    # 转换为灰度图（如果是彩色图）
    if len(image.shape) == 3:
        # 使用RGB转灰度的标准公式
        gray = 0.299 * image[:, :, 0] + 0.587 * image[:, :, 1] + 0.114 * image[:, :, 2]
    else:
        gray = image.copy()
    
    # 确保是uint8格式用于直方图计算
    if gray.max() <= 1.0:
        gray = (gray * 255.0).astype(np.uint8)
    else:
        gray = gray.astype(np.uint8)
    
    # 计算直方图 (256个bins，范围0-255)
    hist, _ = np.histogram(gray.flatten(), bins=256, range=(0, 256))
    
    # 归一化直方图得到概率分布
    hist = hist.astype(np.float64)
    hist_sum = hist.sum()
    if hist_sum > 0:
        hist = hist / hist_sum
    else:
        return 0.0
    
    # 计算信息熵 (避免log(0))
    hist = hist[hist > 0]
    if len(hist) == 0:
        return 0.0
    
    entropy_value = -np.sum(hist * np.log2(hist))
    
    return entropy_value


def calculate_all_metrics(img_ref, img_enhanced):
    """
    计算所有评估指标
    
    Args:
        img_ref: numpy array, 参考图像（原始低光图像或ground truth）
        img_enhanced: numpy array, 增强后的图像
        
    Returns:
        metrics: dict, 包含PSNR、SSIM和熵的字典
    """
    psnr = calculate_psnr(img_ref, img_enhanced)
    ssim = calculate_ssim(img_ref, img_enhanced)
    entropy_value = calculate_entropy(img_enhanced)
    
    return {
        'PSNR': psnr,
        'SSIM': ssim,
        'Entropy': entropy_value
    }

